// src/types.ts
export interface GeneralInfoProps {
    name: string;
    email: string;
    phone: string;
}

export interface EducationProps {
    school: string;
    title: string;
    date: string;
}

export interface ExperienceProps {
    company: string;
    position: string;
    responsibilities: string;
    from: string;
    to: string;
}